refactor: block offset into own file
