+++ src/compaction/tiered.rs
@@ -138,10 +138,10 @@ mod tests {
         file::LEVELS_MANIFEST_FILE,
         level_manifest::LevelManifest,
         segment::{
+            block::offset::BlockOffset,
             block_index::{two_level_index::TwoLevelBlockIndex, BlockIndexImpl},
             file_offsets::FileOffsets,
             meta::{Metadata, SegmentId},
-            value_block::BlockOffset,
             Segment, SegmentInner,
         },
         HashSet, KeyRange, SeqNo,
