+++ src/segment/value_block.rs
@@ -2,38 +2,12 @@
 // This source code is licensed under both the Apache 2.0 and MIT License
 // (found in the LICENSE-* files in the repository)
 
-use super::{block::Block, id::GlobalSegmentId};
+use super::{
+    block::{offset::BlockOffset, Block},
+    id::GlobalSegmentId,
+};
 use crate::{descriptor_table::FileDescriptorTable, value::InternalValue, BlockCache};
 use std::sync::Arc;
-
-#[derive(Copy, Clone, Default, Debug, std::hash::Hash, PartialEq, Eq, Ord, PartialOrd)]
-pub struct BlockOffset(pub u64);
-
-impl std::ops::Deref for BlockOffset {
-    type Target = u64;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl std::ops::AddAssign<Self> for BlockOffset {
-    fn add_assign(&mut self, rhs: Self) {
-        *self += *rhs;
-    }
-}
-
-impl std::ops::AddAssign<u64> for BlockOffset {
-    fn add_assign(&mut self, rhs: u64) {
-        self.0 += rhs;
-    }
-}
-
-impl std::fmt::Display for BlockOffset {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum CachePolicy {
