+++ src/segment/writer/mod.rs
@@ -5,7 +5,7 @@
 mod meta;
 
 use super::{
-    block::header::Header as BlockHeader,
+    block::{header::Header as BlockHeader, offset::BlockOffset},
     block_index::writer::Writer as IndexWriter,
     file_offsets::FileOffsets,
     meta::{CompressionType, Metadata},
@@ -15,7 +15,7 @@ use super::{
 use crate::{
     bloom::BloomFilter,
     file::fsync_directory,
-    segment::{block::ItemSize, value_block::BlockOffset},
+    segment::block::ItemSize,
     value::{InternalValue, UserKey},
     Encode, SegmentId,
 };
